Stat card refactor
